+++ packages/dara-core/changelog.md
@@ -5,6 +5,7 @@ title: Changelog
 ## NEXT
 
 -  Allow passing a `ClientVariable` to the `DynamicComponent` component
+-  Added a new `CopyToClipboard` action (also available as `ctx.copy_to_clipboard`) to copy a given value to the user's clipboard
 
 ## 1.21.21